backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["arturrejment <arturrejment1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.5"
black = "^23.1.0"
isort = "^5.12.0"
djangorestframework-simplejwt = "^5.2.2"
drf-spectacular = "^0.26.1"
django-stubs = "^1.16.0"
mypy = "1.1.1"
djangorestframework-stubs = {extras = ["compatible-mypy"], version = "^1.10.0"}
djangorestframework-camel-case = "^1.4.2"
django-cors-headers = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
